fix: returned topics from query instead of posts and moved to the topics API instead
